net/mlx4_core: Limit min profile numbers
We can't use arbitrarily low profile numbers for resources since the device
needs some of its own. The best way to do it is to query the device and then
set the minimum values, but the values chosen here are good enough and do not
complicate the implementation.

Ported from MLNX_OFED 3.4, originated at 2.4 as:
c8721dee6082 mlx4_core: limit min profile numbers

Squashed into it:
719654dc45f5 mlx4_core: Update minimum size for log_num_qp to 18
And part of:
3cd221f94228 mlx4_core: log_num_mtt handling

And a missing mod_param_profile.num_mtt update from 3.0:
9a0fe0c18b09 mlx4_core: Add missing parts of enable changing HCA resource limits patch

Change-Id: I091b240be29faf57b221d5ff93200c6671ce07cb
Signed-off-by: Eli Cohen <eli@mellanox.co.il>
Signed-off-by: Dotan Barak <dotanb@dev.mellanox.co.il>
Reviewed-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Vladimir Sokolovsky <vlad@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>