IB/mlx4: add blue flame support for kernel consumers
Using blue flame can improve latency by allowing the HW to more efficiently
access the WQE. A consumer who wants to use blue flame, has to create the
QP with inline support. When posting a send WR, the consumer has to set
IB_SEND_INLINE in the send flags. This approach is similar to that take in
userspace; that is, in order to use blue flame you must use inline.
However, if the send WR is too large for blue flame, it will only use
inline. A kernel consumer that creates a QP with inline support, will be
allocated a UAR and a blue flame register. All QP doorbells will be set to
the UAR and blue flame posts to the blue flame register. We make use of all
available registers in a blue flame page.

Also squashed bug fixes from OFED 3.X series:
commit ba3cd0331e69 ("IB/mlx4: Fix create QP error flow")
commit b84799f686cd ("IB/mlx4: deprecate "failed to alloc bf reg" message from err to debug
B/mlx4: deprecate "failed to alloc bf reg" message from err to debug")

issue: 511206
issue: 870322
Change-Id: Ia6dcdd4d6ae33d52942142423a6d451c7063ef09
Signed-off-by: Eli Cohen <eli@mellanox.co.il>
Signed-off-by: Alaa Hleihel <alaa@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Moshe Lazer <moshel@mellanox.com>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>