net/mlx4: Let old type of slaves to work with new hypervisor
Hypervisor that supports the "RoCE Address Table" PRM new command
may get requests to modify the table from a slave thatsupports only
the old command. This patch allows old type of slaves to work with new
hypoervisor as long as the hypervisor is capable oftranslating the old
command to the new command.

issue: 546660
issue: 875512
Change-Id: I33aab64932fed91f4422a5d2712f2b2e8824bb35
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>