{net,IB}/mlx4: Initialize CQ buffers in the driver when possible.
Perform CQ initialization in the driver when the capability is supported
by the FW.  When passing the CQ to HW indicate that the CQ buffer has
been pre-initialized.

Doing so decreases CQ creation time.  Testing on P8 showed a single 2048
entry CQ creation time was reduced to ~170us, from ~395us.

Issue: 547370
Change-Id: I66a2d9ad3985be6b823f8f3e2487761f0cd7624a
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>