IB: Add RSS and TSS QP groups
Took as part of rebase over upstream to support user space EXP RSS API
of CX3 that wasn't accepted upstream.

Note:
Other API was accepted upstream for RSS support.

In general:

RSS (Receive Side Scaling) TSS (Transmit Side Scaling, better known as
MQ/Multi-Queue) are common networking techniques which allow to use
contemporary NICs that support multiple receive and transmit descriptor
queues (multi-queue), see also Documentation/networking/scaling.txt

A QP group is a set of QPs consists of a parent QP and two disjoint sets
of RSS and TSS QPs. The creation of a QP group is a two stage process:

In the the 1st stage, the parent QP is created.
In the 2nd stage the children QPs of the parent are created.

Each child QP indicates if its a RSS or TSS QP. Both the TSS
and RSS sets of QPs should have contiguous QP numbers.

Notes:
1) Before extending ib_exp_qp_init_attr for the new RSS stuff of CX3,
   aligned with upstream ib_qp_init_attr.

2) This patch comes as a pre-patch to expose RSS to user space
applications which uses the EXP API that wasn't accepted upstream.

3) In case IPoIB in the "rebase" mode will need that
   functionality, an extra cap flags should be added on top
   of ib_device_cap_flags in a place which will prevent future
   conflict.

issue: 652143
Change-Id: I5723d315baea6c1acc3c94868ec1ff4d97268ec3
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>