net/mlx4_core: Add interface to disable multicast loopback
Add interface to disable/enable multicast loopback.
This is done by using the SET_PORT command, when
esw_loopback_disabled_mc is set, then the multicast traffic will only
be transmitted to the wire, instead of being also redirected to the
device by loopback.

When multicast loopback is disabled the bandwidth is increased in
RAW QP traffic from 23.97 to 39.85 GB/sec.

Issue: 868009
Change-Id: I8d1e50a67e76d01ced5514281139664d6233c8e3
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>
Signed-off-by: Gal Pressman <galp@mellanox.com>