net/mlx4_core: Extend num_mtt in dev caps to avoid overflow
Some legitimate combinations of log_num_mtt and log_mtts_per_seg cause
overflow in the calculation of the num_mtt when initializing the HCA
which causes Kernel panic. Changed the variable to be 'u64' instead of
'int' to avoid the overflow and made the needed changes to support the new type.

issue: 862533
issue: 190542
Change-Id: I25aecb98003b353c29b64f6835d6742e1078f249
Fixes: 2b8fb2867ca2 ("mlx4_core: mtts resources units changed to offset")
Signed-off-by: Majd Dibbiny <majd@mellanox.com>
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>