IB/mlx4: Enable allocation of DBs and BF pages from user context
Based on Moshe Lazer commit from 3.x.

This patch provides a mechanism to allocate and map UARs and their
related blue-flames(BFs) pages to user space applications.
It uses the mmap mechanism for this allocation.

The patch add 2 new mmap commands:
- MLX4_IB_EXP_MMAP_EXT_UAR_PAGE: to allocate a page of DBs to user
  space.
- MLX4_IB_EXP_MMAP_EXT_BLUE_FLAME_PAGE: to allocate a page of BFs to
  user space.

In addition to the mmap command field the patch provides also
command-parameter.

In this command-parameter the user provides the index of the buffer he
wants to allocate.

The allowed allocation index range is 0 ..MLX4_IB_MAX_CTX_UARS.
The kernel driver keeps linked list of the allocated UARs.

Note:
To allocate page i of BFs, the user must first allocate UAR page i.
This is because BlueFlame pages are affiliated with the UAR pages by
their indexes. A QP can only use a BlueFlame page with the index equal
to the QP UAR.

issue: 862323
Change-Id: Ie423e51eafce878349ef3401c9d6146c7e5fefe0
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>