IB/uverbs: Add EXP verbs prefetching ODP MR pages
On demand paging allows avoiding the pinning of RDMA memory regions' pages.
However, page faults can be expensive and slow down communication
considerably. Sometimes the user program has knowledge that a coming up
communication operation might use a particular part of a memory region. In
these cases we wish to allow the user program to hint the kernel it can go
ahead and prefetch these pages, and map them into the MR, so that the page
fault can be avoided.
This adds a new verb to the uverbs module to support such a call from user
space. It accepts an MR and a memory range to prefetch, and perform roughly
what a pagefault on the same region would have done: use get user pages to
load the requested pages, and map them into the device.
This patch includes the following changes:
 - Add a new uverbs command (ib_uverbs_prefetch_mr).
 - Add a prefetch_mr function to the ib_device struc

Issue: 865148
Change-Id: I5063094459f7ee96d791e8781727701054bb77e6
Signed-off-by: Mark Bloch <markb@mellanox.com>