net/mlx4_core: Support upto 4TB memory regsitration
Current calculation supports upto 1TB. Changing it to 4TB to allow
larger MRs to be used.
MLX4_MAX_LOG_NUM_MTT is limited to 30 rather than 31 since bitmap
operations are using int and not uint.
This code is ported from MLNX_OFED 3.4 and originates in 2.4 in the
following commits:
6d144c2bc0e7 ('net/mlx4_core: the number of MTTs should consider...')
0bf7b1cb82cc ('net/mlx4_core: limit to 4TB of memory registration')

issue: 926102
Change-Id: Ic2023658f48ecf0511136be97cb7506d507b40fb
Signed-off-by: Noa Osherovich <noaos@mellanox.com>