{net,IB}/mlx4: Fix incompatibility with old counter implementation
The old driver implementation of port counters uses 0xFF for the
sink counter. The new implementation uses 0x7f.
In the old implementation, the Hypervisor passed default counter
indices to the guest via QUERY_FUNC_CAP. The new implementation
does not do this (so a VF running the old implementation leaves 0xFF
as its default counter index).

This created a compatibility problem with the new implementation.
For the INIT2RTR QP transition, the VF put 0xFF as the qp's
associated counter index. This index was rejected by the new
implementation Hypervisor (which tested for 0x7F as the sink).
The INIT2RTR transition on the VF then failed, resulting in
the VF driver not working.

The fix for this is to test, in the new implementation, if
the counter index received from the VF is 0xFF. If so, change it
to the new implementation sink counter before proceeding.

A further incompatibility was noticed when there were no more
counters available to allocate. The old implementation returned
the sink counter (0xFF), with a success status. The new implementation
returned its sink counter (0x7f) with a failure status. This failure
return resulted in the VF entering its error flow and entering index 0
for the counter value (incorrect).

The fix for this is to have the new implementation return success as
well -- and for those places where an error return was expected for
the counter allocation (when the sink counter was returned), test
explicitly for the sink counter being returned in addition to testing
for an error return.

issue: 940345
Change-Id: Ia4a3e6271530f0477e51afbf1a7d88f41e5ef300
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>