IB/core : Fix security hole related to client peer invalidation
The flow of context creates (QP,CQ ..)  calls peer_umem_get and
passes 0 as the invalidation context ticket to get_pages. peer_mem
modules does not aware of this behaviour and can invalidate with
ticket 0. This can cause data corruption and data leaks.

The fix is to check in ib_register_peer_memory_client if the
client supplied a valid invalidate callback, if yes then we
set the invalidation_required field. Then in ib_get_peer_client()
if the peer invalidation_required is not set , skip those peers.

issue: 963995

Change-Id: I6c5fe48b7ff631dbf97ac5ae137724b1ba3ddea2
Signed-off-by: Feras Daoud <ferasda@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>