net/mlx5: Add NVMEoF offload type for XRQ object
Note that According to PRM, in case that PAS list in
the WQ indicates on multiple groups of PASs(Like in NVMF),
then each one of them must be aligned to 64 Bytes.
For example, in NVMF offload, the PAS in WQ contains the physical
addresses of the RQ for receiving the non offloaded NVMEoF commands,
followed by the Physical addresses of the staging buffer. The second
list must be aligned to 64 Bytes.

Issue: 920501

Change-Id: I5cd8a40e65d2d3a1ea18071b8377de5aae253ca1
Signed-off-by: Max Gurtovoy <maxg@mellanox.com>