net/mlx5: Limit fw page allocation time
In many cases where the host is low in memory we see that the DMA
mapping start being slower which cause the fw request handling time to
be very long which in turn causes other pending commands (possibly vf commands
which require more fw pages) to timeout.
For this reason we should limit the page request handling time in order
to have enough time to propagate the allocation failure notification to
the pending command gracefully without timeouts.
Naturally this time limit should be derived from the command timeout
time and we chose to make it half command timeout duration.

Issue: 967257
Change-Id: Ia238181378c0c08fd8109a2cbd41ae3a4112ade2
Signed-off-by: Mohamad Haj Yahia <mohamad@mellanox.com>