net/mlx5: Add infrastructure for PDDR register
PDDR (Port Diagnostics Database Register) is used to read the physical
layer debug database, which contains helpful troubleshooting information
regarding the state of the link.

PDDR register can only be queried when PCAM register reports it as
supported in its register mask. A new helper macro was added to
the MLX5_CAP_* infrastructure in order to access this mask.

Issue: 991096
Change-Id: I106205a04d8ba0ca92e8346b5e2f6811c48c6e26
Signed-off-by: Gal Pressman <galp@mellanox.com>