net/mlx5e: Open new channels after deactivating existing ones
In order to WA a kernel bug (soft lockup in IOMMU, under stress),
create new channels only after deactivating the existing ones,
so that there is no simultanuous traffic that might stress the system.
Issue doesn't exist in upstream since kernel v4.7.

This does not break the fail-safe mechanism, it just increases the time
of interface down.

Upstream patches that address and fix the issue:
2aac630429d9 iommu/vt-d: change intel-iommu to use IOVA frame numbers
9257b4a206fc iommu/iova: introduce per-cpu caching to iova allocation
22e2f9fa63b0 iommu/vt-d: Use per-cpu IOVA caching

Issue: 1044389
Change-Id: I9bfeed5bebef39039eb42ee49ab26267a5936349
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>