IB/mlx5: Counter for sent cnaks shows wrong value
The counter cnaks (showed by
/sys/class/infiniband/<dev>/dct/<port>/tx_cnaks) is incremented in
reduce_tx_pending() that is called only when the tx_queue is full. This
mean that for a long time after driver start the counter shows zero and
afterwards is incremented by tx_signal_factor every tx_signal_factor
post_sends().
Move the increase of the counter right after post_send() to fix that.
Note that the meaning of the counter now is slightly changed. Instead of
showing only the successful sends it now shows all sends.

issue: 1094936
Change-Id: Icda7847e709cac73d7448daa92fb18dd56798b0b
Signed-off-by: Moni Shoua <monis@mellanox.com>