IB/ipoib: Fix incorrect SGE number of RSS parent QP
When creating a parent RSS QP, the max SGE was set to 1.
In the past despite this low value a higher value was set when
creating the QP keeping this bug hidden.

The correct behaviour is to set the number of SGE to be the same
as other TSS QPs.

issue: 1120424
Change-Id: Ic3c72b5dde78f105984b43e588b18c3a6645a1c0
Signed-off-by: Alex Vesker <valex@mellanox.com>