net/mlx5e: Add support for configuring pfc stall prevention through sysfs
In the event where the device unexpectedly becomes unresponsive
for a long period of time, flow control mechanism may propagate
pause frames which will cause congestion spreading to the entire
network.
To prevent this scenario, when the device is stalled for a period
longer than a pre-configured timeout, flow control mechanisms are
automatically disabled.

Implement set/get functions to configure PFC stall prevention
mode through sysfs.
On default the stall prevention timeout is configured to 8 sec.
Auto mode will set the stall prevention timeout to be 100 msec.

The pfcc register is used as well to configure the pfc and the
global pause. In order to set the pfcc register, read-modify-write
solution is needed.

In order to get the current pfc stall prevention mode:
$ cat /sys/class/net/ethXX/settings/pfc_stall_prevention
In order to set new pfc stall prevention mode:
$ echo "auto" > /sys/class/net/ethXX/settings/pfc_stall_prevention

Issue: 1041105
Change-Id: Ib3c89d702312d6a6067eed348d2b6c1f4d2e8984
Signe-off-by: Inbar Karmy <inbark@mellanox.com>