net/mlx4_en: Remove high order pre-allocations in RX datapath
Remove the usage of split_page(), as it's unclear what versions
of it exists in the different kernels.
It also complicates the tracking in memtrack.

It was used in the flow of high order pre-allocations,
which is only an optimization, some kind of allocation bulking.

In addition, revert recent change in mlx4_en_recover_from_oom().

Issue: 1161742
Change-Id: Iff9f5bbe995847e38d5cdc95d77c6cb21c28aedc
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>