IB/ipoib: Workaround for ib_notify_cq silently fail race
There is a race on arm_sn value from an event and for arming cq.
This issue is causing ib_notify_cq to silently fail. If arming fails the
QP connected to this CQ will not function causing the queue to be stuck.

The flow that we see in this error during stop:
        Disable NAPI
        Modify QPs -> ERR
        Drain CQs (ib_poll_cq)
        Modify QPs -> RST
		Rearm CQs
        ---> Event due to the QP flush with done by modifying to ERR
        ---> CQ stays in state=0, unarmed

This event causes a race on arm_sn, the arm_CQ doorbell is passed with
an incorrect sequence causing FW to ignore it, leaving CQ unarmed.

issue: 1041786

Change-Id: I8b5078a58e3a3cd38571509cf756a71c83950a03
Signed-off-by: Alex Vesker <valex@mellanox.com>