net/mlx5e: E-Switch, Add peer miss rules
In the sriov offloads mode, packets that are not matched by any
other rule are sent towards the e-switch vport manager for further
processing.

Under upcoming patches (e.g for uplink LAG), packets sent from VF
vports belonging to esw0 (e-switch related to PF0) might end up in
esw1 (e-switch related to PF1) due to muxing logic applied by the FW.

In such a case we still want the missed packet to be sent to the "base"
esw manager vport in order to present the control plane a consistent
view of the source (VF reresentor) port.

The miss rules are initialized for both eswitches when the "last"
(second) eswitch is set to switchdev mode. Since each eswitch cleanup
itself, we also make sure not to reallocate miss rules again if not
needed.

Issue: 1159068
Change-Id: I19990df3bba51f1881b056810166bd05eb8809c1
Signed-off-by: Roi Dayan <roid@mellanox.com>
Signed-off-by: Shahar Klein <shahark@mellanox.com>
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>