IB/mlx5: Support extended number of strides for "Strding RQ"
A new feature was exposed in the device which extends the range of
number of strides that can be used for a single WQE of the Striding
RQ feature. Instead of minimum 512 strides, now it can support minimum
8 strides. Accordingly, min_single_wqe_log_num_of_stride field from
mlx5_ib_exp_query_device will expose the extended range and the
experimental create WQ handler is modified.

Issue: 1250275
Change-Id: I9034df6495287db70b639677b492a3c832181653
Signed-off-by: Guy Levi <guyle@mellanox.com>