IB/umem: Support for rsync invocation in ODP invalidation flow
In Power9 setups, Nvidia GPUs are connected through NVLink, while
the HCA is connected through PCIe. The gateway between them is a NPU.
To achieve optimal performance, the NPU should be set to relaxed
ordering mode, so that reads and writes initiated from the HCA to the
GPU can be executed out of order.
In order to provide consistency, the NPU provides an RSYNC operation
that waits until pending reads/writes have been completed.
This patch adds ODP support to perform Rsync call when a page is
invalidated.
RSYNC operations, control and the enablement of relaxed ordering is
done is a separate module that connects with IB core through special
callbacks.

Issue: 1272095
Change-Id: I2b5bbb14b067d3222e4de4089f41f667678bc584
Signed-off-by: Feras Daoud <ferasda@mellanox.com>