net/mlx5: Handle SW reset of FW in error flow
Some mlx5 adapter FW allows the driver to reset the FW in the event of
an error. When a SW reset is issued on any PF all PFs enter reset state.
This is a recoverable condition. The existing recovery flow was designed
to allow the recovery of a VF after a PF driver reload. This patch expands
the scope of that flow to recover PFs or VFs after a SW reset has been
issued. When a software reset is issued the following occurs:

1. The NIC interface mode is set to 7 while the reset is in progress.
2. Once the reset completes the NIC interface mode is set to 1.

After the reset has been issued (added in a subsequent patch) the health
poll for other functions will detect that the NIC interface state has been
set to disabled. This will cause it to enter the existing recovery flow.
If the PCI is still working (meaning it doesn't return 0xff on all reads)
it means recovery can proceed immediately instead of waiting 60 seconds.

issue: 1187431
Change-Id: I6be8a5fbe42b9c0f7f1575e794c39de74b2bd243
Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>