net/mlx4: Fix bad calculation of port map
When two ports of an HCA in bonded mode are disabled in switch and
when port #2 is being enabled, the virtual to physical port mapping is
miscalculated. Instead of the expected mapping V1 -> P2 and V2 -> P2 we
get V1 -> P1 and V2 -> P2. This is noticed in a virtual function device
that are attached to port #1.
To fix that the driver needs to consider the other port link state when
processing an event on a port instead of making the decision based only
on the info about the state of the port that is associated with the event.

issue: 1275785
Change-Id: I0d0a2b6ded119ab0bfc1ee69356ba2c1891cf2be
Signed-off-by: Moni Shoua <monis@mellanox.com>