net/mlx5: Add cmd_unknown_namespace_cnt to NVMe-oF target offload ctx
This attribute replace offloaded_capsules_count according to PRM
re-design for NVMe-oF target offload counters.
This attribute will be used in QUERY_XRQ cmd, and will count the
number of commands that arrived with unknown/non-existing namespace.
This feature is supported only when cmd_unknown_namespace_namespace_cnt
cap is set.

Issue: 1160848

Change-Id: I38429f5e54bc1b6780f1fbd0a8777409f9aff66d
Signed-off-by: Max Gurtovoy <maxg@mellanox.com>