IB/mlx5: Support extended number of strides for Striding RQ
A new feature was exposed in the device which extends the range of number of
strides that can be used for a single WQE of the Striding RQ feature. Instead of
minimum 512 strides, now it can support minimum 8 strides. Accordingly,
min_single_wqe_log_num_of_strides field from striding_rq_caps will expose the
extended range and the create RQ handler is modified.

issue: 1229798
Change-Id: If3cd20d116806448e32e5a6b6771a52709dddc6e
Signed-off-by: Majd Dibbiny <majd@mellanox.com>