nvmet-rdma: automatic listening port re-activation
In case the device goes away (or resets) we get a device
removal event (or .remove ib_client callback). So what
we want is to destroy the listening cm_id and re-activate
(or enable) when the same device comes back. Hence we introduce
nvmet_rdma_port which stores the ib_device node guid, and when
a new device comes in to the system (ib_client .add callback) we
search for an existing listener port on this device and reconfigure
the listener cm_id.

Issue: 1334850

Change-Id: If62667b74549278416475db318501ea1d8ca2d6f
Signed-off-by: Sagi Grimberg <sagi@grimberg.me>
Signed-off-by: Israel Rukshin <israelr@mellanox.com>