net/mlx5: Reset FW before tearing down driver
If the FW encountered an error but it still operating it may attempt to
access resources the driver destroys during the recovery flow.
Reset FW first, then tear everything down.
This also includes a redesign since flags are no longer needed if
the reset and crdump are done on the same function.

issue: 1187431
Change-Id: I411751e6c6b925a0a6e486106d14609d9ad542e1
Signed-off-by: Alex Vesker <valex@mellanox.com>