net/eipoib: Support bonding (active-passive mode) for eIPoIB driver.
The driver now is able to support bonding interface on top of the pv interfaces
(e.g: ethX and ethX+1).

The main idea is that the bond interface is attached to the bridge; this gives
the High Availability option over the two ports eth0 and eth1 and at the end
over ib0 and ib1.

Whenever one of the physical ports goes down the other available port became
the "Active" and continues the traffic over it.

The implementation uses 2 units, the eIPoIB driver and the user-space application.

The driver does:
1.learns the IP's it serves: via 3 ways:
 a. ARP packet.
 b. packets for miss IP's, which causes ARP request.
 b. from sys/fs.
2.exports these IP's via sys/fs.
3. Whenever new served IP is learned the driver sends gart-arp to the world.

The user-space application now can fetch the served IP from the previous active
interface and restores them in the new active interface.

Issue: 1369461
Change-Id: If5064d0eafbdb63db9b9c2dcd0038b9d58c75fe0
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Denis Drozdov <denisd@mellanox.com>