mlx5/core: capi performance improvement
ODP uses a single in-flight workqueue which can cause a big delay
between when the page fault interrupt happens and when the page fault
workqueue is serviced.

For CAPI page fault, change the page fault workqueue to high priority
workqueue with maximum 16 in-flight workqueue.

This fix can apply for ODP page fault case, but I decide to hold off
until we do more validation.

Issue: 1357043
Change-Id: I9d7cec073333022a50600730f0e4191292ca6f7f
Signed-off-by: Huy Nguyen <huyn@mellanox.com>