mlx5/core: disable CAPI in mlx5_pci_err_detected
Disable CAPI in mlx5_pci_err_detected. This is to disable CAPI during
EEH recovery. Note that CAPI is not enabled in mlx5_pci_resume
because of a following bug in P9. Also note that because of this bug,
reloading the mlx5_core driver after EEH will hit EEH because the driver
initialization has no idea about the previous EEH to avoid
initializaing CAPI.

HW440288 : RXE_ARB Error Injection Broken on P9
"You are hitting a known bug in P9 with this particular injection.
Basically, this injection should only result in an ER error condition to
occur but instead, after cleaning up the hardware, a packet that follows
will cause a timeout. This was found late in the program and there are
no fixes for this. There are probably other ways to achieve what you are
trying to do to avoid this. For instance, using setpci in linux, you
should be able to clear the memory space enable in the adapter's config
space. That should result in the error case you are trying to do here
with the hardware injector."

Issue: 1349681
Change-Id: I58f2d34541d20a2f45a07f2808a4ae3deaa4c6aa
Signed-off-by: Huy Nguyen <huyn@mellanox.com>