ib/ipoib: add support for clones / multiple child interfaces
Allow creating "clone" child interfaces which further partition an
IPoIB interface to sub interfaces who either use the same pkey as
their parent or use the same pkey as already created child interface.

Each child now has a child index, which together with the pkey is
used as the identifier of the created network device.

All sorts of child interfaces are still created/deleted through sysfs,
in a similar manner to the way legacy child interfaces are.

A major use case for clone child is for virtualization purposes, where
a per VM NIC is desired at the hypervisor level, this is simple approach
to the solution provided e.g by means of multi PCI function (Multiple
PF, SR-IOV) and alike.

Issue: 1369461
Change-Id: I226313423797ae5555c6609f9a912820c77f9df8
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Vladimir Sokolovsky <vlad@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Denis Drozdov <denisd@mellanox.com>