ib/ipoib: Add VIF support in IB receive functions
Whenever new IPoIB interface acts as a VIF for some upper
interface/layer, it uses the skb cb storage area to place information
which can be used by driver/layer the uses that VIF.

The eipoib driver, when it uses the IPoIB VIF interface it
needs to generate a source mac for the Ethernet frame.  Handler is
registered and wants to get the data before it goes up to the upper
layer it can get registered over the ipoib module via
netdev_rx_handler_register api, the required data will be sent via the
cb area in the skb.

The changes:
  - According to registered handler calls netif_receive_skb, or
    napi_gro_receive.
  - RSS support for VIFS

Issue: 1369461
Change-Id: I6ff68485e613855b137be4c7f9aa2827bfd711d3
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Denis Drozdov <denisd@mellanox.com>