mlx5/ipoib: Indicate TSS support for enhanced devices
The OFED only implementation for RSS/TSS used for ConnectX3
devices require both sides to support TSS, this support is
indicated using a bit in the hardware address.
This indication is needed due to a possible issue in which a
different source QP numbers will be used for sends from the
same host, this can lead to incorrect mapping on non-Mellanox
devices (para-virtualization).
The source of this problem is due to a HW limitation in QP-groups
which doesn't mask the sender QP.

Although this issue is not present in the enhanced IPoIB mlx5
implementation, there is still an interoperability issue if
datagram mode is used between ConnectX3 devices and enhanced
IPoIB. Setting this bit on mlx5 devices will allow reaching
better performance.

issue: 1400449
Change-Id: I2e22293e9046b680d60e14def0f149aef88f0ec7
Signed-off-by: Alex Vesker <valex@mellanox.com>