net/mlx5: Enable VST with QinQ option by vport_context
Depends on esw capability to support, VST mode behavior is changed to
insert cvlan tag either if there is already vlan tag on the packet or
not. Previous VST mode behavior was to insert cvlan tag only if there
is no vlan tag already on the packet.
This change enables sending packets with two cvlan tags, the outer is
added by the eswitch.

Issue: 1407658
Change-Id: I59af93f4db4e219b071510def718025219251848
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>