IB/ipoib: Allow setting mac address for the VF
Certain orchestation software tries to set the same mac address as what
is currently configured. When such mac address is set, allow setting it
as it is not harmful.

Currently this is not needed in upstream kernel because when old and
new mac address are same, kernel is not attempting to call
unimplemented ndo_set_mac of the NIC driver.

issue: 1417042
Change-Id: I5f6c84b4339ff03c99f3b87ddce809f4e4577518
Signed-off-by: Parav Pandit <parav@mellanox.com>