IB/uverbs: Manipulate DC INI AH fields for backward compatibility
When modifying a DC INI from INIT to RTR, only the SL is valid.  Since old
applications set zeros or invalid values for the rest of the fields, we might
fail down the road in rdma_check_ah_attr.
Change those values to avoid failures and backward compatibility issues.

issue: 1429179
Change-Id: Ib26f00abaebe8bcb2f4d68007e1d5d7b7e0b4966
Signed-off-by: Majd Dibbiny <majd@mellanox.com>