mlx5/core: Use multi threaded workqueue for page fault handling
Page fault events are processed in a workqueue context. Instead of
queueing work items, each is a handler for a single page fault event, to
a single threaded workqueue do it with a multi threaded workqueue and
increase the level of concurrency for of handlers. The number of threads
was chosen as the number of channels to the command interface.
A test showed that before this change 85% of page faults were waiting in
queue 8 seconds or more while after the change 98% of page faults were
waiting in queue 64 milliseconds or less.

issue: 1424907
Signed-off-by: Moni Shoua <monis@mellanox.com>

Conflicts:
	metadata/Moni_Shoua.csv

Change-Id: I4678fd620ab51e66455cdd8135ec429925282263