IB/mlx5: Use fragmented SRQ's buffer for in-kernel users
User may request large SRQ size which may be failed due to low memory
resources on the host. Previously, the driver asked for a continuously
memory which increased the probability that the memory allocation would
be failed and then the user request.
Accordingly, the driver is taking advantage on mlx5_core new feature
which lets to allocate fragmented buffer, and applies it on in-kernel user
SRQ creation requests.

Issue: 1480206
Change-Id: I38104afcf09ad7a0eef39c3a2de693bc8f61436d
Signed-off-by: Guy Levi <guyle@mellanox.com>