IB/mlx5: Fix SQ wraparound for kernel QPs
For some WRs and QPTs, the SQ edge wasn't updated after moving the segment
pointer. This will cause memory corruption while writing the inline/ptr data
segments.

Add an update for the SQ edge prior to writing the first inline/ptr data segment
to be on the safe side.

issue: 1489498
Change-Id: I6568da68007d9c08dd43a2eb7d3beabc91181d4c
Signed-off-by: Majd Dibbiny <majd@mellanox.com>