IB/mlx5: Fix fragmented kernel QP in reset state
New fragmented SQ scheme has a dynamic edge point which is updated every
fragment replacement.
Accordingly, the edge should be reinitialized once it is moved to reset
state. Otherwise, it may be an irrelevant edge.

This bug's symptoms may be shown by dumps of error CQEs in dmesg and
memory overrun.

Issue: 1489498
Fixes: 4cc11049b3a8 ("IB/mlx5: Use fragmented QP's buffer for in-kernel users")
Change-Id: Id103088cb7d1e9efda04e2c289456810f496514c
Signed-off-by: Guy Levi <guyle@mellanox.com>