IB/mlx5: Fix fragmented kernel QP post REG_MR WR
KLMs list of UMR control segment of send WQE should be aligned to 64B.
It should be counted by data segments filed of the control segment but the
code aligned it just after the data segment was set.

This bug's symptoms may be shown by dumps of error CQEs in dmesg and
memory overrun.

Issue: 1489963
Change-Id: I71beaedde132048b4baf8df87e93dd5974b02f80
Fixes: 4cc11049b3a8 ("IB/mlx5: Use fragmented QP's buffer for in-kernel users")
Signed-off-by: Guy Levi <guyle@mellanox.com>