net/sched: Fix incorrect match on vlan packet in centos 7.2
To support centos7.2 with 3.10 kernel, we add a workround in openvswitch
to pass rule with ALL protocol, instead of the specified protocol, to tc
flower. A vlan packet then will be matched incorrectly on a rule for
untagged packets.
To avoid this, we need to check if the rule is set with vlan mask for a
vlan packet.

Issue: 1348065
Change-Id: I71158093aeeffd42a6402a8337c8ad429697dfdf
Signed-off-by: Jianbo Liu <jianbol@mellanox.com>