IB/mlx5: Support device memory type attribute
This patch intoruduces a new mlx5 specific attribute
to the DM allocation method - the DM type.

In order to allow addition of new types in downstream patches
this patch also refactors the allocation, deallocation and
registration handlers to consider the requested type and
perform the necessary actions according to it.

Issue: 1437850
Change-Id: Id40bb72ba622bb68d81019d38e45b5b5b8876b15
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>