RDMA/mlx5: ingress_log_flow_table_size should be part of the flow DB
Ingress_data was added to the mlx5_ib device and not to the flow db,
in addition the stage to initialize it was added just to the uplink
representor. Which means when trying to create flows on the IB
representors device we will access zeroed out values as the flow
table size, which means no flows can be added.

Move steering_data to the flow db, make the uplink representor initialize
it, the store/show function are provided only on the uplink representor
device as it controls those values.

Issue: 1540073
Change-Id: I0828426f7f73c889ac1ebd2d01357d4a912ad1e9
Fixes: ("IB/mlx5: Add ingress_log_flow_table_size sysfs")
Signed-off-by: Mark Bloch <markb@mellanox.com>