net/mlx5e: Expand XPS cpumask to cover all online cpus
Currently we have one cpu in XPS cpumask per tx queue, this is good
enough for default configuration where there is a tx queue per cpu.
However, once configuration changes to use less tx queues, part of the
cpus are not XPS-mapped and so the select queue decision falls back to
hash calculation and balancing is not guaranteed.

Expand XPS cpumask to enable using all cpus even when number of tx
queues is smaller than number of cpus.

Issue: 1523555
Issue: 1531817
Change-Id: Ied8c8ea516a119aa69ae8aeef464ef354d193015
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>