IB/mlx5: Save actual memic allocated size in the mlx5 object
When allocating memic using ibv_exp_alloc_dm API the driver code
didn't store the actual allocated size in the mlx5_ib_dm object
which caused the deallocation not to clear the corresponding page(s)
index bits in the allocation bitmap so they didn't appear as free
pages after the deallocation.

The patch fixes this by storing the actual allocated size in the
object.

Issue: 1548878
Change-Id: I529159383a2ae6997e2f759c947ac92b8b5e3b0d
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>