mlx5/core: Get pp rate limit queues based on capability
There are two customers cases reported about long boot time because of
rate limit queues. That is because driver needs to register large
number of special queues(e.g, 512) for each netdev. And this happens
at boot time which is handled by udev. If the number of udev children
are limited, this will takes significant amount of time to boot.

This patch addressed the issue above by checking packet pacing
capability when registering netdev. Special queue will be set to 0
if packet pacing is not supported. This way works because packet pacing
is disabled as default in firmware, and enable it requires update
mlxconfig and reset fw/driver.

Issue: 1296094
Issue: 1153317
Issue: 919545
Change-Id: I2130cb7b5313727b156a0e1d3f2fa1ad125e1a66
Signed-off-by: Bodong Wang <bodong@mellanox.com>
Signed-off-by: Yuval Avnery <yuvalav@mellanox.com>