IB/mlx5: Fix SW ICM bitmaps size calculation
Adding consideration of the ICM start address into the ICM
blocks bitmap size calculation. If address is zero it means
the memory is not supported by the device.

The previous calculation didn't consider the case where
FW doesn't support an ICM type which caused the size calculation
to produce a very large number and a memory allocation
failure.

Issue: 1548518
Change-Id: I6aedd69fa99c4e080194de26718a3bf05ebf81b7
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>