net/mlx5e: Enlarge TX inline header size in case of IP mode to include IPv4/IPv6 header
In case SKB metadata for transport offset is too small to cover both L2
and IP headers (i.e. false function indication), the driver forces the
copy of a size that guarantees covering any L2 and IP headers, including
a VLAN.

This workaround is set in order to solve kernel bugs in which the
skb_transport_offset was set wrongly.

Issue: 1554108
Change-Id: I790b4a415d2870a06f4e594363590bbf6c600d48
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>