net/mlx5e: Handle port mtu/link, dcb and lag for uplink reps
Take care of setup/teardown for the port link, dcb, lag as well as
dealing with port mtu and carrier for e-switch uplink representors.

This is achieved by adding a dedicated profile instance for uplink
representors which includes the enable/disable and more profile routines
which are invoked by the general mlx5e code for netdev attach/detach.

[upstream commit b36cdb42ad4b]

Issue: 1681087
Change-Id: I26802584d2831ea814ae4dd5ab38f8a5592963df
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>