RDMA/core: Delete RoCE GID in hw when corresponding IP is deleted
Currently a RoCE GID entry is removed from the hardware when all
references to the GID entry drop to zero. This is incorrect.
GID entry should be removed from the hardware when GID entry deletion
is requested.
This allows to terminate ongoing traffic through such RoCE GID.

While a GID is deleted from the hardware, GID slot in the software GID
cache is not freed. GID slot is freed once all references of such GID
are dropped.
This continue to ensure that such GID slot of hardware is not allocated
to new GID entry allocation request. It is allocated once all references
to GID entry drop.

issue : 961226
Change-Id: I276b31d7989bbc3a865e28449ea4db7efd713f6e
Fixes: b150c3862d21 ("IB/core: Introduce GID entry reference counts")
Signed-off-by: Parav Pandit <parav@mellanox.com>