net/mlx5: Eswitch, Rate limiting per VF group
Adding support to apply TX rate limiting on a group of VF.

The individual rate limit per VF will still take affect but
will be bounded by the limitation on the group, if such exists.

When supported by device, all VFs start by default at group 0
and can be moved to other groups by writing to the dedicated new
'group' attribute in the VF's sysfs.

A new 'groups' sysfs is added under the 'sriov' hierarchy which show
the existing groups and allows user to modify the TX rate limit of the
group.

Issue: 1597954
Change-Id: I55abe373fc5bb2930e1145e7746e033d536efd83
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>