nvmet-rdma: fix offload ctrl creation condition
Currently we create the offload ctrl on the first connect command,
but this is based on the assumption that the initiator will not
send the second connect command in case the first command will fail.
This assumption is not always true and depends on initiator
implementation. Change the condition and create the offload ctrl in
case it wasn't already created, regardless of the qid. Also check if
the offload ctrl exsist priorer disabling offload namespace.

Issue: 1564040

Change-Id: I6195f979ec06f1a421991520df875d5aad2ac803
Signed-off-by: Max Gurtovoy <maxg@mellanox.com>
Signed-off-by: Israel Rukshin <israelr@mellanox.com>