net/mlx5: Pass flow steering objects to fs_cmd
Pass the flow steering objects instead of their
attributes to fs_cmd in order to decrease number
of arguments and in addition it will be used to
update object fields.
In addition pass the flow steering root namespace
instead of the device so will have context to the
namespace in the fs_cmd layer.

Issue: 1640818
Change-Id: Ibee1a8fbd2bd05c79255ffc7a36fd38fe74020c3
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>