RDMA/mlx5: Allow inserting a steering rule to the FDB
Allow this only via mlx5 raw create flow API, legacy verbs are not
supported. To accommodate that, we add a new attribute to matcher creation
to indicate the type of flow table to be used.
	MLX5_IB_ATTR_FLOW_MATCHER_FT_TYPE
With this new attribute MLX5_IB_ATTR_FLOW_MATCHER_FLOW_FLAGS is no
longer need, we keep it for computability but at most only a single
attribute can be passed of the two.

When inserting a flow rule to the FDB we require that a DEVX FT
is provided as a destination, no other configuration is allowed.

Issue: 1668005
Change-Id: Iccbb011f7b1246022d4e23a022767e1c3319259c
Signed-off-by: Mark Bloch <markb@mellanox.com>