RDMA/mlx5: Move to single device multiport ports in switchdev mode
Move from IB device (representor) per virtual function to single IB device
with port per virtual function (port 1 represents the uplink). As number
of ports is a static property of an IB device, declare the IB device with
as many port as the possible according to the PCI bus.

Issue: 1668005
Change-Id: If051dc20743cc42aea874d7011fa115bd6698883
Signed-off-by: Mark Bloch <markb@mellanox.com>