IB/mlx5: Consider vlan of lower netdev for macvlan GID entries
When a given netdev of the GID entry is macvlan netdevice, and if the
lower netdevice is vlan device, GID entry for macvlan based IP address
needs to inherit the vlan of the lower netdevice.

Therefore, attempt to find out if the lower device exist and if so, if
it is vlan device and setup the vlan tag correctly.

issue: 1649264
Change-Id: I8ba66914c239e288ffcc915ea154505280abf5fb
Signed-off-by: Yuval Avnery <yuvalav@mellanox.com>