IB/mlx5: Allocate the per-port Q counter shared when DEVX is  supported
The per-port Q counter is some kernel resource and as such may be used
by few UID(s) upon DEVX usage.

To enable using it for QP/RQ when DEVX context is used need to allocate
it with a sharing mode indication to let firmware allows its usage.

The UID = 0xffff was chosen to mark it.

issue: 1656788
Change-Id: I52f35266eeb84c3cb2eb47792b06aa8e981df866
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>