IB/mlx5: Adapt upstream to MLNX_OFED locking scheme
Below patch [1] from 2016 is not upstream ? in upstream we
use  rwlock_t while in OFED we override it with spinlock_t lock.

Need to send the below upstream or drop it at all, for now adapting DEVX
mkey to MLNX_OFED locking scheme.

NOTE:
In case nothing will be done need in next rebase to squash this patch
into the below patch that changed the locking scheme.

[1]
commit 9f4a9b8ec7d389ae67687585c2c81c86e61073ef
Author: Mark Bloch <markb@mellanox.com>
Date:   Wed Dec 7 11:04:07 2016 +0000

    IB/mlx5: Change locking and index on mkey radix tree

Fixes: 9f4a9b8ec7d3 ("IB/mlx5: Change locking and index on mkey radix tree")
issue: 1656790
Change-Id: I3417b44763d90df5cc14f19e1aab7e4ca29b7239
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>