BACKPORTS: Add more checking for nd_tbl in ipv6_stub
In kernel 4.9 and later, a global ipv6_stub data is defined and its nd_tbl
is NULL. The ipv6_stub pointer we are using is pointing to it on first
initialization. It could be reassigned to point to ipv6_stub_impl in ipv6
initialization function.

However, this assignment will not happen if we disable ipv6. For example,
adding ipv6.disable=1 in the kernel boot params. In this case, kernel will
panic when changing to eswitchdev mode because ipv6_stub->nd_tbl is NULL.

To avoid this, we need to check ipv6_stub and its nd_tbl to make sure both
of them are not NULL.

Issue: 1599658
Change-Id: I410faa88b3b5f5d95f2327da6becb2c0f1c161ed
Signed-off-by: Jianbo Liu <jianbol@mellanox.com>