BACKPORTS: Fix agent refcount in core mad.c backport
The case where HAVE_IDR_LOCK was not defined had a
misplaced '}' in function find_mad_agent.

As a result, the agent ref-count was not incremented when
response MADs were received.  This resulted in the agent
ref-count becoming negative -- and causing hangs when the
agent was unregistered (complete was not called, and
wait_for_completion) waited forever.

issue: 1698740
issue: 1663361
Change-Id: If7050003eb0ab3b0968ede3f471266b672622c0e
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>