net/mlx5: E-Switch, fix peer miss rule NULL pointer panic
When disabling SRIOV, system panic because the peer miss rule is NULL.
Compared with upstream code, remove the following functions
	esw_del_fdb_peer_miss_rules(esw);
	mlx5_destroy_flow_group(esw->fdb_table.offloads.peer_miss_grp);
in esw_offloads_steering_cleanup().

I think it is a backport error. In upstream, peer miss rule change is
integrated first and then devcom/VF LAG. In ofed 4.6, the order is
reversed.

Issue: 1705092
Change-Id: I743218ac1c34aa0062b605a4ea87550173ccd861
Signed-off-by: Chris Mi <chrism@mellanox.com>