mlx5/core: Use fast teardown in the driver remove flow
The current implementation of the "device remove callback" synchronise
resources cleanup with the FW. For each resource, it sends destroy
command to the FW and waits for its completion. As a result, it takes
more than 1 second to unbind one PF.

The following change comes to improve this by calling fast teardown before
cleaning the resources. This will lead to the following:
- FLR will be sent to the FW
- Device state moves to internal error

Then the driver will continue with his regular cleanup without the need to
wait for FW completions. This changes decreased the remove time to be less
than 100ms per PF.

cherry-picked from 4.5-5
commit 6dc943b581fc ("mlx5/core: Use fast teardown in the driver remove flow")

Issue: 1598519
Change-Id: Ia0764725831306dbf7b9a4ec502ed694d8d07d5d
Signed-off-by: Feras Daoud <ferasda@mellanox.com>