net/mlx5e: Extract hairpin entry lookup/create code into function
To remove dependency on rtnl lock, code that looks up hairpin entry and
creates new one if entry is not found, must do this atomically to be safe
when concurrent updates are possible. Extract code that looks up or creates
new hairpin entry into standalone function as a preparation for described
refactoring.

Issue: 1715919
Change-Id: I41bbabbaad4639e81162b2e73a44e9ade1c8e605
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>
Reviewed-by: Jianbo Liu <jianbol@mellanox.com>