net/mlx5e: Extract mod hdr entry lookup/create code into function
To remove dependency on rtnl lock, code that looks up mod header entry and
creates new one if entry is not found, must do this atomically to be safe
when concurrent updates are possible. Extract code that looks up or creates
new mod header entry into standalone function as a preparation for
described refactoring.

Issue: 1715919
Change-Id: Ib11a66d25ada51f0d54993c438f6b1f5edb54d2b
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>
Reviewed-by: Jianbo Liu <jianbol@mellanox.com>