net/mlx5e: Extract encap entry lookup/create code into function
To remove dependency on rtnl lock, code that looks up encap entry and
creates new one if entry is not found, must do this atomically to be safe
when concurrent updates are possible. Extract code that looks up or creates
new mod header entry into standalone function as a preparation for
described refactoring.

Issue: 1715919
Change-Id: Ib38934842930bc2144b0c9d10e2b20a974f0a643
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>
Reviewed-by: Jianbo Liu <jianbol@mellanox.com>