IB/mlx5: Fix scatter to CQE in DCT QP creation
The scatter 2 CQE over DCT QP actually has never been enabled since the
driver refers to device QP context layout instead of device DCT context
layout when it is configure the scatter 2 cqe relevant bits. That
latter is used as an object which goes down to device and is parsed
there as a DCT layout.

Issue: 1683936
Fixes: 5d6ff1babe ("IB/mlx5: Support scatter to CQE for DC transport type")
Change-Id: I61ec2cbef471c7e5a4d40ff3696e315a26680a82
Signed-off-by: Guy Levi <guyle@mellanox.com>