net/mlx5e: Fix a race between update_stats and q_counters destroy
Update stats requires q_counters statistics which are provided by the FW.
The q_counters are created in the attach procedure and destroyed in the
detached procedure. Therefore, before updating stats and rescheduling
another update stats delayed work, verify that the netdev device is
still attached by testing that the destroying bit is cleared.

Issue: 1710787
Change-Id: I3c4e631e0314121408c55de87675626709aa49ae
Signed-off-by: Aya Levin <ayal@mellanox.com>