BACKPORTS: ib_core: Fix addr.c copy_src_l2_addr() backport
When macro HAVE_RT_USES_GATEWAY is NOT defined, the backport adds an
argument to procedure has_gateway(). The added argument is a pointer to
the destination IP address.

This is done properly when has_gateway is called from procedure fetch_ha()
(also in file addr.c).

However, in procedure copy_src_l2_addr() the backport was incorrect.
It called procedure has_gateway() with the destination L2 (MAC) structure,
instead of passing the destination IP address.

As a result, has_gateway() incorrectly returned TRUE instead of FALSE.
When there is a gateway, we require RoCE_v2 (routable RoCE), so connection
requests were (incorrectly) rejected when the listener was running RoCE_V1
and there was no gateway in the request packet route.

issue: 1710008
Change-Id: I92962bc4fde27e2e8a19f56008564f390c087246
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>