IB/mlx5: Expose a capability for scat2cqe over DCT QP
The bug which was fixed in commit a06303605a04 ("IB/mlx5: Fix scatter to
CQE in DCT QP creation"), could also cause a memory corruption. To avoid
it, a new capability is exposed to user-space which lets it to ensure
that it will not enable the feature without the fix.

Issue: 1683936
Change-Id: I06086ed85ac83be2ef816ea0a44d205c4780885d
Fixes: a06303605a04 ("IB/mlx5: Fix scatter to CQE in DCT QP creation")
Signed-off-by: Guy Levi <guyle@mellanox.com>