net/mlx5e: Use flags MLX5E_TC_ESW_OFFLOAD to choose the right hashtable
The offending commit is backported from kernel 4.20.16. In kernel 4.20.16,
there is no uplink representor, so We can use esw->mode to to choose the
right hashtable, switchdev mode hashtable or legacy mode hashtable.

But after we support uplink representor, we should use flags got from
block callback. When enabling switchdev mode, the netdev could be PF
or uplink representor. So we can't use esw->mode.

In the following BACKPORTS patch, we need to fix centos 7.2 because
there is no block callback.

Issue: 1718998
Change-Id: I6fa1e370d95e131b853a0daaf3b4993788db845a
Fixes: e1dde43a1f42 ("net/mlx5e: Protect tc flows hashtable with rcu")
Signed-off-by: Chris Mi <chrism@mellanox.com>